Divide page elements into widgets.
