+++ lib/presentation/page/home.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sample_flutter_game_with_flame/presentation/page/drawer.dart';
 import 'package:sample_flutter_game_with_flame/presentation/page/game.dart';
+import 'package:sample_flutter_game_with_flame/presentation/widget/drawer/go_to_home_button.dart';
 
 class HomePage extends ConsumerWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -22,14 +23,7 @@ class HomePage extends ConsumerWidget {
           padding: const EdgeInsets.all(30),
           children: <Widget>[
             const Icon(Icons.home, size: 200),
-            ElevatedButton(
-              onPressed: () => Navigator.of(context).push(
-                MaterialPageRoute<void>(
-                  builder: (_) => GamePage(),
-                ),
-              ),
-              child: const Text("go to game"),
-            ),
+            GoToHomeButton(),
           ],
         ),
       ),
