+++ lib/presentation/widget/drawer/go_to_home_button.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:sample_flutter_game_with_flame/presentation/page/home.dart';
+
+class GoToHomeButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return TextButton(
+      onPressed: () {
+        Navigator.of(context).pushAndRemoveUntil(
+            MaterialPageRoute<void>(builder: (_) => const HomePage()),
+            (_) => false);
+      },
+      child: const ListTile(
+        title: Center(child: Text("go to home")),
+        tileColor: Colors.lightGreen,
+        trailing: Icon(Icons.home),
+      ),
+    );
+  }
+}
